requirements.txt
mcp>=0.9.0
azure-monitor-query>=1.3.0
azure-identity>=1.15.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-resource>=23.0.0
